# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Flunq.io Workflow API
  description: |
    REST API for the Flunq.io serverless workflow engine.
    
    This API provides endpoints for managing workflows, executions, and monitoring workflow state.
    All workflows follow the Serverless Workflow Definition Language specification.
  version: 1.0.0
  contact:
    name: Flunq.io API Support
    url: https://github.com/flunq-io/flunq.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.flunq.io/v1
    description: Production server
paths:
...
  /executions:
    get:
      summary: List executions
      description: Retrieve a list of workflow executions
      operationId: listExecutions
      tags:
        - Executions
      parameters:
        - name: workflow_id
          in: query
          description: Filter executions by workflow ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter executions by status
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: limit
          in: query
          description: Maximum number of executions to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of executions to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /executions/{executionId}:
    get:
      summary: Get execution details
      description: Retrieve detailed information about a specific execution
      operationId: getExecution
      tags:
        - Executions
      parameters:
        - name: executionId
          in: path
          required: true
          description: Unique identifier of the execution
          schema:
            type: string
            pattern: '^ex_[a-zA-Z0-9]{16}$'
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Cancel execution
      description: Cancel a running workflow execution
      operationId: cancelExecution
      tags:
        - Executions
      parameters:
        - name: executionId
          in: path
          required: true
          description: Unique identifier of the execution
          schema:
            type: string
            pattern: '^ex_[a-zA-Z0-9]{16}$'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
                  example: "User requested cancellation"
      responses:
        '200':
          description: Execution cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    ExecutionResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the execution
          example: "ex_01234567890abcdef"
        workflow_id:
          type: string
          description: ID of the workflow being executed
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
          description: Current status of the execution
        correlation_id:
          type: string
          description: Correlation ID for tracking
        started_at:
          type: string
          format: date-time
          description: When the execution started
        completed_at:
          type: string
          format: date-time
          description: When the execution completed (if applicable)
    ExecutionDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ExecutionResponse'
        - type: object
          properties:
            input:
              type: object
              description: Input data provided to the execution
              additionalProperties: true
            output:
              type: object
              description: Output data from the execution (if completed)
              additionalProperties: true
            error:
              type: object
              description: Error information (if failed)
              properties:
                message:
                  type: string
                code:
                  type: string
                details:
                  type: object
                  additionalProperties: true
            current_state:
              type: string
              description: Current state in the workflow
            duration_ms:
              type: integer
              description: Execution duration in milliseconds
    ExecutionListResponse:
      type: object
      properties:
        executions:
          type: array
          items:
            $ref: '#/components/schemas/ExecutionResponse'
        total:
          type: integer
          description: Total number of executions
        limit:
          type: integer
          description: Maximum number of executions returned
        offset:
          type: integer
          description: Number of executions skipped
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "The request body contains invalid data"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        request_id:
          type: string
          description: Unique identifier for the request
          example: "req_01234567890abcdef"
  responses:
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              value:
                error: "VALIDATION_ERROR"
                message: "The request body contains invalid data"
                details:
                  field: "name"
                  reason: "name is required"
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            workflow_not_found:
              value:
                error: "WORKFLOW_NOT_FOUND"
                message: "The specified workflow does not exist"
    InternalServerError:
      description: Internal Server Error - Something went wrong
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal_error:
              value:
                error: "INTERNAL_ERROR"
                message: "An unexpected error occurred"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication (to be implemented)

# Global security (commented out for now)
# security:
#   - BearerAuth: []
...